ChangeResult: use variants to for now
This is a temporary hack to make ChangeResult usable for a 2D table
view. It requires QPointF and QSizeF instead of qreals to represent
position and size changes, respectively. Later on, we should find a
way to do this without heap allocation in ChangeResult (QVariant) or
in layout() (polymorphic ChangeResult).

Change-Id: I53ffdd622fd8ff81b02b0a9e6592295fb56fc8d3
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>