AbstractFxViewItem: add new class: AbstractFxViewItem
QQuickItemView makes use of a wrapper class around a
a QQuickItem called FxViewItem. This class holds some
extra information related to the item, like any ongoing
transition it might be undergoing.

Since such a class is also needed for QQuickTableView, factor
the common parts out to AbstractFxViewItem, and let
FxViewItem inherit from it.

Change-Id: I8db7fc1cb138097c2326f54749431a2e55488535