AbstractFxViewItem: add new class: AbstractFxViewItem
QQuickItemView makes use of a wrapper class around a
a QQuickItem called FxViewItem. This class holds some
extra information related to the item, like any ongoing
transition it might be undergoing.

Since such a class is also needed for QQuickTableView, factor
the common parts out to AbstractFxViewItem, and let
FxViewItem inherit from it.

Change-Id: Id9b9c6f4de144d3a9594d8162e2a807cd3a564da