QQmlAdaptorModel: make a models row/column count read-only
This change partly revert 59a9c7c3d9edeb9, since we no
longer make use of the introduced API.

Being able to override the row/column count for a delegate model
was a wrong step. If a view needs to operate on a row/column count
that is different from what the model offers, then the mapping
should be done in the view (or in a proxy model), and not in the
delegate model.

Change-Id: I32b0dfa977dd7cae33c399e138aac847e49aa94a