QQmlDelegateModel: implement support for delegate item recycling
This patch will make it possible for the views (TableView) to
recycle delegate items.

The added code will not affect any existing views (or usage of
DelegateModel) as it stands. Recycling of items needs to be enabled
by the view, and for now, this will only be done by TableView.

Recycling of items is specified per item. So when the view tells the
model to release an item, it can specify whether or not it's recyclable.
If the view never releases any items as recyclable (which will be the
case for ListView and GridView), no items will be added to the recycle
pool, and hence, no items will be recycled.

If the view releases an item as recyclable, the model will move it
to the recycle pool rather than destroying it (if the items ref-count
is zero, and not persisted). The next time the view asks the model for
a new item, the model will first check if the recycle pool already
contains an item, and if so, use it. Otherwise it will create a new
item, like before.

Items in the pool are supposed to rest there for a short while. Ideally
only from the time items are flicked out on one side of the view, until
we reuse them when new items are flicked in on the oppsite side. One big
reason for this is that we have no way of hibernating items in QML, so
they will effectively stay alive while inside the pool. This is not
ideal, but still a huge performance improvement what we got from before,
where all items are always created from scratch. If hibernating objects
will ever be possible, it should be easy to extends the current logic to
take advantage of this.

Already existing tests (for e.g ListView etc) should verify that no
regressions are introduced with this patch. Since recycling of items
is driven from the view, auto tests for the new API is included
with the patch for implementing recycling support in TableView (coming
in a subsequent patch).

Change-Id: Iaa2865e3b6fdee2b1a2dbd060c46c3da4982d19e