TableView: switch to use implicit width/height for delegates
This is logically more correct since TableView will override
width / height anyway to make the delegates fit into the table.
And it gets even more problematic when we recycle delegates, since
in that case we need to keep the original width binding of the
delegate to calulate the size of new columns. And this all
fits better with using implicitWidth/height.

Change-Id: Ia5f2acd2bfc45f3fb160c3782191ad8da9f780e6