TableView: support delegate item recycling
This patch will make use of the new API in QQmlDelegateModel
to support delegate item recycling.

The API to control this will mainly be a new property
"recycleItems", which is true by default. By setting it to
false, no recycling will ever happen.

In addition, each delegate item will have an attached property
"TableView.recyclable", which is also true by default. By setting
it, the user can control recycling on a per-item level, granted
that recycling is already enabled on the table view (recycleItems).

When an item is recycled, the signal "TableView.recycled" is
emitted after the fact, in case the delegate item needs to
execute some extra code during the process.

From an implementation point of view, TableView only needs to
do two things to enable recycling. First, whenever it releases
items, it will add a second argument to release(), informing
QQmlDelegateModel if the item can be recycled. Second, it
needs to call drainRecyclePool() at appropriate times to ensure
that no items will be kept alive in the pool for too long.

Change-Id: I830e2eace776302ac58946733566208aa8954159