TableView: load and unload edges in the same loop
If you flick the table really fast so that the whole table
ends up outside the screen, the current logic would remove all
edges from the table, then start builing the table up again, edge
by edge (until the whole viewport was filled up). The downside of
this strategy is that we can easily end up loading more edges than
what actually ends up visible, since we need to rebuild the table
from where we left it after flicking, possibly far outside the
viewport. This edges will be removed once the table is rebuild, but
in the meantime, we'll keep the hidden edges in memory.

This patch will change the logic a bit so that we unload an edge, then
load an edge, and repeat this until the table is filled up. That way
we don't keep hidden edges in memory longer than necessary.

Change-Id: Ia0f1968a4b3579e4445e1f7b6e68a28a1d2b360b