QQmlAdaptorModel: add function to test if an adaptor model supports recycling
This patch will prepare QQmlAdaptorModel for delegate item recycling.
It basically provides a way to query if a model adaptor supports recycling
of delegate items.

To begin with, recycling will only be supported for QAIM models. For that
adapter, functionality already exists for notifying the application
about changes to the role values (which will be needed when recycling an
item, since then all the role values for the item will change).

More patches will follow to enable recycling for other model adaptors as well.

Change-Id: I56d7cadd85cf75b335f746ae31cc6fcc911a4ca4