QQmlAdaptorModel: support item recycling for js list models
This patch will enable delegate item recycling for js list models.
So if you e.g set "model: [1, 2, 3, 4]" on a TableView, the delegates
can be recycled.

The patch will override the notify function. This function is called from
QQmlDelegateModel whenever the modelData that the delegate item
depends on changes. This again is needed to trigger all the modelData
bindings in the item to update.

Auto testing this is already handled in
tst_QQuickTableView::checkAttachedProperties(). This test will pick up
that the list adaptor now supports recycling, and check that the delegate
items still contain the correct context properties.

Change-Id: I49185bdbaaacc3ccbd94c99cc66d9a1998452f68