QQmlAdaptorModel: support item recycling for QQmlListModels
This patch will enable delegate item recycling for QQmlListModels.
So if you e.g set "model: someItem.children" on a TableView, the delegates
can be recycled.

The patch will override the notify function. This function is called from
QQmlDelegateModel whenever the modelData that the delegate item
depends on changes. This again is needed to trigger all the modelData
bindings in the item to update.

Auto testing this is already handled in
tst_QQuickTableView::checkAttachedProperties(). This test will pick up
that the list adaptor now supports recycling, and check that the delegate
items still contain the correct context properties.

Change-Id: I34161a998a4581798b692da69164ef29032d6e4f