QQuickTableView: preload one extra row and column at start-up
When a TableView initially loads as many rows and columns it can fit
inside the viewport, it will always be one less than the number it
will show while flicking. The reason is that, as soon as you flick
half a column out on the left, half a column will move in on the right.
And this will increase the number of visible columns by 1 (but without
reusing any items from the pool, since the first column is not out).

Since this is always the case, it makes sense to preload one extra
row and column at start-up, so that they're ready when the flicking
starts.

Note that this doesn't load more items in the background than
what we need (like the cache buffer would). The viewport will fit
_all_ the loaded items into the viewport once you start flicking.
But the extra items loaded at start-up will instead be moved direcly
to the pool for reuse, and the application will be informed about it
(using the onPooled signal).

Change-Id: Icea85c1d44f74ab54f1b96325489e8d6d1c0889e