QQmlDelegateModel: guard new row/column properties with revision 12
Tag the new 'row' and 'column' properties with the correct
revision for 5.12. This will make sure that they cannot be
accessed by the delegate unless the QQmlAdaptorModel has the
correct minorVersion.

Setting the minor version will currently only be done by
QQuickTableView, which means that row and column can only be
accessed from delegates used with TableView, and not e.g
ListView or Repeater.

Fixes: QTBUG-70031
Change-Id: I49e67c37ab5b7925c7bca313bbb99f04d1387cc4