QQmlPropertyCache: support setting an explicit minorVersion
When creating a QQmlPropertyCache for a QMetaObject, there were
currently no way to specify which version to use. Normally this
is not needed, since when creating property caches for types
declared in QML, the version would be filled in later, based on
the import version found in the QML file.

But sometimes we need to create a QQmlPropertyCache for a QMetaObject
created in C++, that has no associated QML file and import version.
And if that meta object has revisioned properties, we need to
specify which version of the meta object the cache should represent.
Otherwise, the version would just be 0, which means that revisoned
properties would not be found by the V4 runtime later.

As an example, QQmlAdaptorModel has a set of classes that wraps various
models (QAIM, arrays, etc). When a new delegate item is created by a
view, an instance of a model class will be created as well (from C++).
This instance will be set as context object for the delegate
item, enabling properties such as index, row, column and model roles.
But since row and column should a revision (currently they
don't), we need to be able to specify that the property cache should
have the same version as the view. That way, we can ensure that they
don't shadow any existing row and column properties that might exist in
the application from before, and as such, cause regressions.

This patch will add an extra argument to the constructor that lets
you specify the version of the QMetaObject.

Task-number: QTBUG-70031
Change-Id: I1c245a0c8b6f071e35865966fedc97f2839cd2f3