QQmlAdaptorModel::Accessors: create a propertyCache for all accessors, not just for QAIM
From before, only accessors for wrapping a QAbstractItemModel
had to create a dynamic QMetaObject together with a shared
QQmlPropertyCache (for enabling model roles in the delegate).
Each model item in the view would get a QQmlData with the dynamic
property cache assigned, which would then later be used by the
v4 runtime during property lookup.

But after we added the properties 'row' and 'column' to the model items, we
now always need to create a property cache, regardless of the Accessor used.
That way we can we specify the correct version of the model item in the
cache, which will also allow us to revision the new properties so that they
will be respected by the v4 runtime.

This patch will move the 'metaObject' and 'propertyCache' up to the base
class (Accessor), and ensure that we create a property cache for each of the
non-pure-virtual sub classes. The model item wrappers will then, when creating
a QQmlData, assign the shared cache from the associated Accessor.

Finally, we add a small API to QQmlObjectModel (from which QQmlAdaptorModel
and QQmlTableInstanceModel derive), that lets you specify which minor version
the model items should have.

Task-number: QTBUG-70031
Change-Id: If6a67d5968d360d4a2b23d8291669c0549e8a342