QQuickTableView: add new 'syncView' property
This property can be set to point to another TableView.
If set, this TableView will be synchronized to the
other table with respect to flicking, column width, row
heights, spacing, etc. This logic is needed as a foundation
for the upcoming HeaderView.

Upcoming patches will implement this logic (together with
autotests) gradually.

Change-Id: Ic7dea8e1d1aa46bbb3ea6e795953a65c96c25cc6