QQuickTableView: implement recursive updateTable()
Now that a TableView can be inside a syncView hierarchy, we
cannot update a table in isolation, but need to coordinate
this with the other views. It's especially important that
we update a parent syncView before a child syncView, to
ensure that the parent has calculated all the necessary
columns width and row heights. For that reason, we always
update the table views starting from the top.

Change-Id: Iba8ae7d28fa0bb2fbbad9f8fc7aa198e15b91872