QQuickTableView: factor out scheduleRebuildIfNeededAfterViewportMoved()
Factor most of the existing code in 'viewportMoved' into a separate
function, since we have to add more (slightly unrelated) code into
'viewportMoved' later.

At the same time, we also update the logic to check if the table has
been flicked back into the viewport. When the view is synced with another
view, and that view has a model with more columns than us, it can happen
that our table ends up being flicked completly out of view. If that
happens, we detect when we're flicked back in, and do a rebuild to ensure
that we're still in sync.

Change-Id: I5a5b8e795426484eeab3771f6c8d4c9b7da046eb