QQuickTableView: update viewportMoved() to take syncView into account
Now that several table views can stay in sync through the
syncView parent-child chain, we also need to ensure that the
position of the content views stays in sync. This patch will
recursively go through all connected views when one of the
views are moved and set the same position on them all according
to the syncDirection flag.

Change-Id: I5a5b8e795426484eeab3771f6c8d4c9b7da046eb