QQuickTableView: check if we need to rebuild when syncing with syncView
Since different tables can have different sized
models, it can also happen, as the views are being flicked
around, that some views temporarily end up with no
visible rows and columns in the viewport. When that
happens, we continually check if the columns should
become visible again, and if so, schedule a rebuild.

Change-Id: I5a5b8e795426484eeab3771f6c8d4c9b7da046eb