QQuickTableView: change enforceTableAtOrigin() to instead rebuild table
enforceTableAtOrigin() takes care of checking if the table
is not aligned with the edges of the content view, and if not, move
the items to the correct place. This can sometimes happen if you fast
flick around and cause rebuilds to happen. This is meant as
a temporary solution until we have better understanding how we can
instead (if possible) manipulate originX/Y in QQuickFlickable.

Since we can no longer just move the rows and column in a table
without also ensuring that the other connected tables stay in
sync, this patch will change the implementation of this function
to fix that. The easiest way to ensure this, is to just rebuild
the table when the situation occur (since that will cause the other
tables to rebuild as well).

Change-Id: If88340fee39fba12d931e1542b80e9ed518bb477