see if I can avoid recalculate content width
Another approach is to never calculate table endExtent, but instead
modify contentWidth. So I move origin, but resize the contents.

The reason why that doesn't always work, is that the app can set a
custom contentWidth. And I should still anyway keep the edges aligned.

Perhaps I should modify origin but not extents/contentWidth if a custom
contentWidth is set. So if the table is not right aligned,

Change-Id: I786a964e8dbf256beed0d1ad4c028aa11cc0aecc