QQuickTableView: improve updateAverageEdgeSize()
When we know the exact size of the content view, we can take advantage
of this to calculate the exact average cell size. This in turn will
improve the positioning of rows and columns whenever we need to
rebuild, since we have a better idea where they should end up in
the content view.

Change-Id: I46c3e87eb38ab032df7c28b6144d1b2de1b9d4ef